Create nuget packages
