+++ .gitignore
@@ -1,4 +1,8 @@
 *.userprefs
 *.suo
+*.user
 output/
 obj/
+.idea/
+.vs/
+.vscode/
