+++ CHANGELOG.md
@@ -1,0 +1,21 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
+and this project adheres to [Semantic Versioning](http://semver.org/).
+
+<!-- Available types of changes:
+### Added
+### Changed
+### Fixed
+### Deprecated
+### Removed
+### Security
+-->
+
+## [Unreleased]
+
+### Changed
+
+- Create nuget package
